requirements.txt
pytest

plaid-python
python-decouple
